refactor(users): reorganize infrastructure layer
This commit moves certain files to the infrastructure layer of the 'users' module, aligning better with their functionality.
